refactor: 컨벤션에 맞게 파일명을 kebab-case로 작성
