+++ src/components/common/DropDown/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import DropDownIcon from '@assets/icons/dropdown_btn.svg';
+import DropDownIcon from '@assets/icons/dropdown-btn.svg';
 import { UseFormRegisterReturn } from 'react-hook-form';
 import { css } from 'styled-system/css';
